Use %lu with (unsigned long) casts for uid_t and gid_t
POSIX only guarantees uid_t/gid_t are integer types, not specifically
unsigned int. Cast to (unsigned long) with %lu to avoid truncation on
platforms where these types are wider than unsigned int. Also updates
the two existing (unsigned)/%u usages for consistency.

+++ root.c
@@ -341,7 +341,7 @@ void ensure_permitted(void)
             error("You must be in the %s group to run root", groupname);
         }
         else {
-            error("You must be in group %u to run root", (unsigned)ROOT_GID);
+            error("You must be in group %lu to run root", (unsigned long)ROOT_GID);
         }
         exit(ROOT_PERMISSION_DENIED);
     }
